Clean-up and more on EntityTagElement
- former 'height' variaible re-named 'entityBlocksHigh'
- former 'width' variable re-named 'entityBlocksWidth'

For clarity, intermediate variables introduced:
- blockScale = scale * BLOCK_HEIGHT
- entityHeight = blockScale * entityBlocksHigh
- entityWidth = blockScale * entityBlocksWide

- render() now uses local getEntityOffset(), in the manner <img> and <item> now use Alignable's getOffset().
This was motivated as renderEntityFollowsMouse takes it's 'x' parameter as the _center_ of the entity.

- local @override of getHeight() expressed with entityBlockHeight and blockScale

- Constant BLOCK_HEIGHT changed to 24, for this reasoning:
As currently rendered by Hermes, text lines are 10 tall (8 for letters, plus 2 of spacing);
therefore 5 lines of text are 48 tall: 5 * 8 tall lines, plus 4 * 2 tall spacing in between.
At BLOCK_HEIGHT = 24, a two-block-high entity will set nicely against 5 lines of text (when side-by-side using
<column>s).
- The literal integer offset in getHeight() was changed to 2, as part of the same reasoning.

I'll be re-evaluating this reasoning by surveying the vanilla entities in <entity> tags.
Commiting now to use as a baseline for that review/survey.